+++ front_end/src/pages/IDPWLogin.js
@@ -1,9 +1,7 @@
-import React, { useState } from "react";
 import styled from "styled-components";
-import { useNavigate } from "react-router-dom";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
-// import { Link } from 'react-router-dom';
+import useLoginCheck from "../hooks/api/useLoginCheck";
 
 const Container = styled.div`
   display: flex;
@@ -81,41 +79,17 @@ const AdditionalLinks = styled.div`
 // `;
 
 const IDPWLogin = () => {
-  const [userData, setUserData] = useState({
-    userid: "",
-    userpwd: "",
-  });
-  const navigate = useNavigate();
+  // 커스텀 훅스
+  const { userData, setUserData, loginCheck } = useLoginCheck();
 
-  // Login Submit
-  const handleLogin = async (event) => {
+  const handleLogin = (event) => {
     event.preventDefault();
     if (Object.values(userData).some((value) => value === "")) {
       alert("아이디와 비밀번호를 입력해주세요");
       return;
     }
-
-    try {
-      const res = await fetch("/api/loginCheck", {
-        method: "POST",
-        body: JSON.stringify(userData),
-        credentials: "include",
-        headers: { "Content-Type": "application/json" },
-      });
-      const data = await res.json();
-
-      alert(data);
-      if (res.status === 200) {
-        navigate("/");
-      } else {
-        setUserData({ userid: "", userpwd: "" });
-        return;
-      }
-    } catch (err) {
-      console.log(err);
-    }
+    loginCheck();
   };
-
   const handleChange = (e) => {
     const { name, value } = e.target;
     setUserData({ ...userData, [name]: value });
@@ -123,83 +97,39 @@ const IDPWLogin = () => {
 
   return (
     <>
-    <Header />
-    <Container>
-      <LoginForm onSubmit={handleLogin}>
-        <H2>로그인</H2>
-        <FormGroup>
-          <Label>아이디</Label>
-          <InputField
-            type="text"
-            name="userid"
-            value={userData.userid}
-            onChange={handleChange}
-            placeholder="아이디를 입력하세요"
-          />
-        </FormGroup>
-        <FormGroup>
-          <Label>패스워드</Label>
-          <InputField
-            type="password"
-            name="userpwd"
-            value={userData.userpwd}
-            onChange={handleChange}
-            placeholder="비밀번호를 입력하세요"
-          />
-        </FormGroup>
-        <SubmitButton type="submit">로그인</SubmitButton>
-        {/* <AdditionalLinks>
+      <Header />
+      <Container>
+        <LoginForm onSubmit={handleLogin}>
+          <H2>로그인</H2>
+          <FormGroup>
+            <Label>아이디</Label>
+            <InputField
+              type="text"
+              name="userid"
+              value={userData.userid}
+              onChange={handleChange}
+              placeholder="아이디를 입력하세요"
+            />
+          </FormGroup>
+          <FormGroup>
+            <Label>패스워드</Label>
+            <InputField
+              type="password"
+              name="userpwd"
+              value={userData.userpwd}
+              onChange={handleChange}
+              placeholder="비밀번호를 입력하세요"
+            />
+          </FormGroup>
+          <SubmitButton type="submit">로그인</SubmitButton>
+          {/* <AdditionalLinks>
                     <AdditionalLink to="/">아이디 / 비밀번호 찾기</AdditionalLink>
                     <AdditionalLink to="/">회원가입</AdditionalLink>
                 </AdditionalLinks> */}
-      </LoginForm>
-    </Container>
-    <Footer />
+        </LoginForm>
+      </Container>
+      <Footer />
     </>
   );
 };
-
-//================ 기존코드 ============================//
-// const IDPWLogin = () => {
-//     const [username, setUsername] = useState('');
-//     const [password, setPassword] = useState('');
-
-//     const handleLogin = (e) => {
-//         e.preventDefault();
-//         console.log('Username:', username);
-//         console.log('Password:', password);
-//     };
-
-//     return (
-//         <Container>
-//             <LoginForm onSubmit={handleLogin}>
-//                 <H2>로그인</H2>
-//                 <FormGroup>
-//                     <Label>아이디</Label>
-//                     <InputField
-//                         type="text"
-//                         placeholder="아이디를 입력하세요"
-//                         value={username}
-//                         onChange={(e) => setUsername(e.target.value)}
-//                     />
-//                 </FormGroup>
-//                 <FormGroup>
-//                     <Label>패스워드</Label>
-//                     <InputField
-//                         type="password"
-//                         placeholder="비밀번호를 입력하세요"
-//                         value={password}
-//                         onChange={(e) => setPassword(e.target.value)}
-//                     />
-//                 </FormGroup>
-//                 <SubmitButton type="submit">로그인</SubmitButton>
-//                 {/* <AdditionalLinks>
-//                     <AdditionalLink to="/">아이디 / 비밀번호 찾기</AdditionalLink>
-//                     <AdditionalLink to="/">회원가입</AdditionalLink>
-//                 </AdditionalLinks> */}
-//             </LoginForm>
-//         </Container>
-//     );
-// };
-
 export default IDPWLogin;
